Refactor attendance list and delete functionality
Renames StudentList.html to AttendanceList.html,

Update the attendance list UI and delete workflow to use AJAX and confirmation buttons.

Updates related JS and CSS for improved interactivity and styling.

Refactors backend delete view to not require a class ID as we now work on the attendance record ID in the post request, and now returns a 404 if the attendance record is not found [This needs improvement],

Updates templates and URLs to use the new attendance list and delete endpoints.

Add a referer_path template filter.  This is used in the attendance record delete flow to provide a "next" parameter, even if it's not technically needed as we do the delete via AJAX.

Some JS changes replacing double quotes with single quotes in line with similar changes in the Python code.