Centralise date formatting and XHR functions, improve attendance list formatting, introduce if-modified-since to attendance list polling.
Create a global GLOBGOR object for global utility functions - macchiato.js - this is loaded by mainlayout.html so should be included in all views.

Centralise date / time client side formatting into GLOBGOR and expand functionality,
Centralise XHR functions into GLOBGOR

Improve use of data attributes when formatting dates and times:
Use data-isodate to indicate that the client should format the provided date time as a date.
Use data-isotime to indicate that the client should format the provided date time as a time.
Use date-isoopt to provide custom locale options via data-localeopts

Improve attendance list formatting by using template context to explicitly enable date / class headers instead of relying on weird CSS rules.  headers are hidden by default and are overridden to always be shown for the export form view.  The context variables are passed into the data-querypath attribute in the attendance list template so are passed on for update polling requests.

Keep track of the last-modified time when the attendance list is updated.  Update times are tracked per class and per module.  Pass the update time to the client via the standard last-modified header and use the if-modified-since header in the request in addition to if-none-match to limit database queries and hashing on the server side.
> Introduce a LocalState class and STATE constant to track this local state.