Update paths, update qr code generation
Generate QR code dynamically when requested directly instead of generating it when the view is requested.

Use content negotiation to pick the preferred image format - This will now default to whatever the browser wants when requested through an <img> tag, or SVG if browsed to directly.

Allow forcing the format via extension or mime type.

Update paths to better semantically separate faculty and student views.